On branch cleanup  Changes to be committed: 	modified:   .gitignore 	deleted:    org.flatpak.EVETerm.yml 	deleted:    pyvenv.cfg 	deleted:    requirements.txt 	renamed:    auth.py -> src/auth.py 	renamed:    config.py -> src/config.py 	renamed:    esi_oauth_native.py -> src/esi_oauth_native.py 	renamed:    esi_oauth_web.py -> src/esi_oauth_web.py 	renamed:    main.py -> src/main.py 	renamed:    mods.py -> src/mods.py 	renamed:    revoke_refresh_token.py -> src/revoke_refresh_token.py 	renamed:    shared_flow.py -> src/shared_flow.py 	renamed:    validate_jwt.py -> src/validate_jwt.py
cleaned up the hellish structure. I hope it runs smooth.
