--- org.flatpak.EVETerm.yml
@@ -1,13 +1,0 @@
-app-id: org.flatpak.EVETerm
-runtime: org.freedesktop.Platform
-runtime-version: '19.08'
-sdk: org.freedesktop.Sdk
-command: python3 main.py
-modules:
-  - name: evetem
-    buildsystem: simple
-    build-commands:
-      - install -D hello.sh /app/bin/hello.sh
-    sources:
-      - type: file
-        path: hello.sh
